Fully qualified publics in verilator control files

+++ src/V3Dead.cpp
@@ -418,6 +418,7 @@ class DeadVisitor final : public VNVisitor {
     }
     bool mightElimVar(const AstVar* nodep) const {
         if (nodep->isSigPublic()) return false;  // Can't elim publics!
+        if (nodep->mayBecomePublic()) return false;  // .vlt -path'ed candidate
         if (nodep->isIO() || nodep->isClassMember() || nodep->sensIfacep()) return false;
         if (nodep->isTemp() && !nodep->isTrace()) return true;
         return m_elimUserVars;  // Post-Trace can kill most anything
